Moved DefaultPlayerOperationsTestableBuilder to separated class
