+++ src/test/java/testing/stub/NullDeviceOperations.java
@@ -1,0 +1,46 @@
+package testing.stub;
+
+import com.odeyalo.sonata.connect.model.CurrentPlayerState;
+import com.odeyalo.sonata.connect.model.Device;
+import com.odeyalo.sonata.connect.model.Devices;
+import com.odeyalo.sonata.connect.model.User;
+import com.odeyalo.sonata.connect.service.player.DeviceOperations;
+import com.odeyalo.sonata.connect.service.player.DisconnectDeviceArgs;
+import com.odeyalo.sonata.connect.service.player.SwitchDeviceCommandArgs;
+import com.odeyalo.sonata.connect.service.player.TargetDeactivationDevices;
+import com.odeyalo.sonata.connect.service.player.sync.TargetDevices;
+import org.jetbrains.annotations.NotNull;
+import reactor.core.publisher.Mono;
+
+public class NullDeviceOperations implements DeviceOperations {
+
+    @NotNull
+    @Override
+    public Mono<CurrentPlayerState> addDevice(User user, Device device) {
+        return Mono.empty();
+    }
+
+    @NotNull
+    @Override
+    public Mono<Boolean> containsById(User user, String deviceId) {
+        return Mono.empty();
+    }
+
+    @NotNull
+    @Override
+    public Mono<CurrentPlayerState> transferPlayback(User user, SwitchDeviceCommandArgs args, TargetDeactivationDevices deactivationDevices, TargetDevices targetDevices) {
+        return Mono.empty();
+    }
+
+    @NotNull
+    @Override
+    public Mono<CurrentPlayerState> disconnectDevice(User user, DisconnectDeviceArgs args) {
+        return Mono.empty();
+    }
+
+    @NotNull
+    @Override
+    public Mono<Devices> getConnectedDevices(User user) {
+        return Mono.empty();
+    }
+}
